Warning fixes for very aggressive gcc settings.
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@84 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/xmlrpc-c/src/cpptest.cc
@@ -215,7 +215,7 @@ void test_value (void) {
     TEST(strct.structSize() == 2);
     TEST(strct.structHasKey("bar"));
     TEST(!strct.structHasKey("nosuch"));
-    for (int i = 0; i < strct.structSize(); i++) {
+    for (size_t i = 0; i < strct.structSize(); i++) {
 	string key;
 	XmlRpcValue value;
 	strct.structGetKeyAndValue(i, key, value);
